+++ Task1/jmp-parent/jmp-bank-api/src/main/java/com/epam/bank/api/App.java
@@ -1,4 +1,4 @@
-package com.epam.jmp;
+package com.epam.bank.api;
 
 /**
  * Hello world!
