+++ Task1/jmp-parent/jmp-bank-api/src/main/java/com/epam/bank/api/service/Bank.java
@@ -1,4 +1,4 @@
-package com.epam.jmp.service;
+package com.epam.bank.api.service;
 
 import com.epam.jmp.dto.BankCard;
 import com.epam.jmp.dto.BankCardType;
