+++ Task1/jmp-parent/jmp-bank-api/src/main/java/module-info.java
@@ -1,5 +1,5 @@
 module jmp.bank.api {
     requires jmp.dto;
 
-    exports com.epam.jmp.service;
+    exports com.epam.bank.api.service;
 }
