+++ Task1/jmp-parent/jmp-cloud-bank-impl/src/main/java/com/epam/jmp/bank/App.java
@@ -1,4 +1,4 @@
-package com.epam.jmp;
+package com.epam.jmp.bank;
 
 /**
  * Hello world!
