+++ Task1/jmp-parent/jmp-cloud-bank-impl/src/main/java/module-info.java
@@ -2,5 +2,5 @@ module jmp.cloud.bank.impl {
     requires transitive jmp.bank.api;
     requires jmp.dto;
 
-    exports com.epam.jmp.impl;
+    exports com.epam.jmp.bank.impl;
 }
