+++ Task1/jmp-parent/jmp-cloud-bank-impl/src/test/java/com/epam/jmp/bank/AppTest.java
@@ -1,4 +1,4 @@
-package com.epam.jmp;
+package com.epam.jmp.bank;
 
 import junit.framework.Test;
 import junit.framework.TestCase;
