+++ Task1/jmp-parent/jmp-cloud-service-impl/src/main/java/com/epam/jmp/service/App.java
@@ -1,4 +1,4 @@
-package com.epam.jmp;
+package com.epam.jmp.service;
 
 /**
  * Hello world!
