+++ Task1/jmp-parent/jmp-cloud-service-impl/src/main/java/module-info.java
@@ -2,5 +2,5 @@ module jmp.cloud.service.impl {
     requires transitive jmp.service.api;
     requires jmp.dto;
 
-    exports com.epam.jmp.impl;
+    exports com.epam.jmp.service.impl;
 }
