+++ Task1/jmp-parent/jmp-cloud-service-impl/src/test/java/com/epam/jmp/service/AppTest.java
@@ -1,4 +1,4 @@
-package com.epam.jmp;
+package com.epam.jmp.service;
 
 import junit.framework.Test;
 import junit.framework.TestCase;
