+++ Task1/jmp-parent/jmp-dto/src/main/java/module-info.java
@@ -1,0 +1,3 @@
+module jmp.dto {
+    exports com.epam.jmp.dto;
+}
