+++ Task1/jmp-parent/jmp-service-api/src/main/java/com/epam/jmp/service/api/App.java
@@ -1,4 +1,4 @@
-package com.epam.jmp;
+package com.epam.jmp.service.api;
 
 /**
  * Hello world!
