+++ Task1/jmp-parent/jmp-service-api/src/main/java/com/epam/jmp/service/api/service/Service.java
@@ -1,4 +1,4 @@
-package com.epam.jmp.service;
+package com.epam.jmp.service.api.service;
 
 import com.epam.jmp.dto.BankCard;
 import com.epam.jmp.dto.Subscription;
