+++ Task1/jmp-parent/jmp-service-api/src/main/java/module-info.java
@@ -1,5 +1,5 @@
 module jmp.service.api {
     requires jmp.dto;
 
-    exports com.epam.jmp.service;
+    exports com.epam.jmp.service.api.service;
 }
