+++ Task1/jmp-parent/jmp-service-api/src/test/java/com/epam/jmp/service/api/AppTest.java
@@ -1,4 +1,4 @@
-package com.epam.jmp;
+package com.epam.jmp.service.api;
 
 import junit.framework.Test;
 import junit.framework.TestCase;
